requirements.txt
async-timeout==4.0.3
blinker==1.7.0
Brotli==1.1.0
click==8.1.7
cssmin==0.2.0
deepmerge==1.1.1
Deprecated==1.2.14
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.1
Flask==3.0.2
Flask-Assets==2.1.0
Flask-Compress==1.14
Flask-Limiter==3.5.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
govuk-frontend-jinja==3.0.0
govuk-frontend-wtf==3.0.0
gunicorn==21.2.0
idna==3.6
importlib-resources==6.1.1
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsmin==3.0.1
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
pluggy==1.5.0
Pygments==2.17.2
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.1
rich==13.7.0
tomli==2.0.1
typing_extensions==4.9.0
webassets==2.0
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.2
zipp==3.18.1
